Add a VFXPlatform 2017 (ish) test case
There's no osl-ci:2017 container, but that's fine. If we use the
osl-ci:2019 container and build just a couple of packages on our own,
the deviations are too minor to worry about.

The important problem this solves is that somehow we'd lapsed into
not having any of our automated CI tests try to build with gcc 4.8.

Signed-off-by: Larry Gritz <lg@larrygritz.com>