Remove direct access to ShaderGlobals members from llvm_ops
This is a stepping stone to "shader globals placement".

Turns out that there were only two functions in llvm_ops.cpp that
directly needed to know the layout of the ShaderGlobals struct: the
implementations of osl_calculatenormal, and osl_raytype_bit.

For osl_calculatenormal, just pass the flipHandedness flag directly,
instead of the ShaderGlobals ptr, from which the field was extracted.

For osl_raytype_bit, similar, and in fact, no implementation is needed
in llvm_ops at all, it's so simple that I just made llvm_gen_raytype
fully generate the necessary bitcode.

Signed-off-by: Larry Gritz <lg@larrygritz.com>