Optional texture argument for colorspace (part 1 -- don't reject)
A while back, we proposed a new optional token/value parameter for the
texture functions: "colorspace", followed by a color space name, which
indicates that the texture name referenced is understood to be in that
colorspace (this is passed to the TextureSystem, which has known for
some time how to ensure the right transformations are applied as tile
data is read in).

We do not fully implement this yet! It's more involved than I have
time to get right between now and the scheduled OSL 1.14 non-beta
relase on Apr 1. Not wanting to delay things, this change at least
causes the use of the new parameter (which apparently, MaterialX has
been using) to be silently accepted and ignored, rather than
generating runtime errors. We'll come back right away for a more
complete working implementation.

Signed-off-by: Larry Gritz <lg@larrygritz.com>